requirements.txt
py==1.4.34
pycryptodomex==3.4.6
pyperclip==1.5.27
pytest==3.1.2
GitPython==3.1.7
